+++ tests/Slots/InventoryStackSlot/InventoryStackSlotTests.try_replace.cs
@@ -18,8 +18,8 @@ namespace TheChest.Inventories.Tests.Slots.InventoryStackSlot
         {
             var stackSize = this.random.Next(MIN_STACK_SIZE_TEST, MAX_STACK_SIZE_TEST);
             var slot = this.slotFactory.Empty(stackSize);
-            var items = new T[] { default! };
 
+            var items = new T[] { default! };
             Assert.That(() => slot.TryReplace(items, out _), Throws.ArgumentNullException);
         }
 
@@ -29,8 +29,8 @@ namespace TheChest.Inventories.Tests.Slots.InventoryStackSlot
             var stackSize = this.random.Next(MIN_STACK_SIZE_TEST, MAX_STACK_SIZE_TEST);
             var slotItems = this.itemFactory.CreateMany(stackSize);
             var slot = this.slotFactory.WithItems(slotItems, stackSize);
-            var items = Array.Empty<T>();
 
+            var items = Array.Empty<T>();
             var result = slot.TryReplace(items, out _);
 
             Assert.That(result, Is.False);
@@ -42,8 +42,8 @@ namespace TheChest.Inventories.Tests.Slots.InventoryStackSlot
             var stackSize = this.random.Next(MIN_STACK_SIZE_TEST, MAX_STACK_SIZE_TEST);
             var slotItems = this.itemFactory.CreateMany(stackSize);
             var slot = this.slotFactory.WithItems(slotItems, stackSize);
-            var items = Array.Empty<T>();
 
+            var items = Array.Empty<T>();
             slot.TryReplace(items, out var oldItems);
 
             Assert.That(oldItems, Is.Null);
@@ -55,8 +55,8 @@ namespace TheChest.Inventories.Tests.Slots.InventoryStackSlot
             var stackSize = this.random.Next(MIN_STACK_SIZE_TEST, MAX_STACK_SIZE_TEST);
             var slotItems = this.itemFactory.CreateMany(stackSize);
             var slot = this.slotFactory.WithItems(slotItems, stackSize);
-            var items = Array.Empty<T>();
 
+            var items = Array.Empty<T>();
             slot.TryReplace(items, out _);
 
             Assert.That(slot.GetContents(), Is.EqualTo(slotItems));
@@ -67,8 +67,8 @@ namespace TheChest.Inventories.Tests.Slots.InventoryStackSlot
         {
             var stackSize = this.random.Next(MIN_STACK_SIZE_TEST, MAX_STACK_SIZE_TEST);
             var slot = this.slotFactory.Empty(stackSize);
-            var items = this.itemFactory.CreateMany(stackSize + 1);
 
+            var items = this.itemFactory.CreateMany(stackSize + 1);
             var result = slot.TryReplace(items, out _);
 
             Assert.That(result, Is.False);
@@ -79,8 +79,8 @@ namespace TheChest.Inventories.Tests.Slots.InventoryStackSlot
         {
             var stackSize = this.random.Next(MIN_STACK_SIZE_TEST, MAX_STACK_SIZE_TEST);
             var slot = this.slotFactory.Empty(stackSize);
-            var items = this.itemFactory.CreateMany(stackSize + 1);
 
+            var items = this.itemFactory.CreateMany(stackSize + 1);
             slot.TryReplace(items, out var oldItems);
 
             Assert.That(oldItems, Is.Null);
@@ -91,8 +91,8 @@ namespace TheChest.Inventories.Tests.Slots.InventoryStackSlot
         {
             var stackSize = this.random.Next(MIN_STACK_SIZE_TEST, MAX_STACK_SIZE_TEST);
             var slot = this.slotFactory.Empty(stackSize);
-            var items = new[] { this.itemFactory.CreateDefault(), this.itemFactory.CreateRandom() };
 
+            var items = new[] { this.itemFactory.CreateDefault(), this.itemFactory.CreateRandom() };
             var result = slot.TryReplace(items, out _);
 
             Assert.That(result, Is.False);
@@ -103,8 +103,8 @@ namespace TheChest.Inventories.Tests.Slots.InventoryStackSlot
         {
             var stackSize = this.random.Next(MIN_STACK_SIZE_TEST, MAX_STACK_SIZE_TEST);
             var slot = this.slotFactory.Empty(stackSize);
-            var items = this.itemFactory.CreateMany(this.random.Next(1, stackSize));
 
+            var items = this.itemFactory.CreateMany(this.random.Next(1, stackSize));
             var result = slot.TryReplace(items, out _);
 
             Assert.That(result, Is.True);
@@ -115,8 +115,8 @@ namespace TheChest.Inventories.Tests.Slots.InventoryStackSlot
         {
             var stackSize = this.random.Next(MIN_STACK_SIZE_TEST, MAX_STACK_SIZE_TEST);
             var slot = this.slotFactory.Empty(stackSize);
-            var items = this.itemFactory.CreateMany(this.random.Next(1, stackSize));
 
+            var items = this.itemFactory.CreateMany(this.random.Next(1, stackSize));
             slot.TryReplace(items, out var oldItems);
 
             Assert.That(oldItems, Is.Empty);
@@ -127,9 +127,9 @@ namespace TheChest.Inventories.Tests.Slots.InventoryStackSlot
         {
             var stackSize = this.random.Next(MIN_STACK_SIZE_TEST, MAX_STACK_SIZE_TEST);
             var slot = this.slotFactory.Empty(stackSize);
+
             var items = this.itemFactory.CreateMany(this.random.Next(1, stackSize));
             var expectedItems = (T[])items.Clone();
-
             slot.TryReplace(items, out _);
 
             Assert.That(slot.GetContents(), Is.EqualTo(expectedItems));
@@ -141,8 +141,8 @@ namespace TheChest.Inventories.Tests.Slots.InventoryStackSlot
             var stackSize = this.random.Next(MIN_STACK_SIZE_TEST, MAX_STACK_SIZE_TEST);
             var slotItems = this.itemFactory.CreateMany(stackSize);
             var slot = this.slotFactory.WithItems(slotItems, stackSize);
-            var replacingItems = this.itemFactory.CreateManyRandom(this.random.Next(1, stackSize));
 
+            var replacingItems = this.itemFactory.CreateManyRandom(this.random.Next(1, stackSize));
             var result = slot.TryReplace(replacingItems, out _);
 
             Assert.That(result, Is.True);
@@ -154,8 +154,8 @@ namespace TheChest.Inventories.Tests.Slots.InventoryStackSlot
             var stackSize = this.random.Next(MIN_STACK_SIZE_TEST, MAX_STACK_SIZE_TEST);
             var slotItems = this.itemFactory.CreateMany(stackSize);
             var slot = this.slotFactory.WithItems(slotItems, stackSize);
-            var replacingItems = this.itemFactory.CreateManyRandom(this.random.Next(1, stackSize));
 
+            var replacingItems = this.itemFactory.CreateManyRandom(this.random.Next(1, stackSize));
             slot.TryReplace(replacingItems, out var oldItems);
 
             Assert.That(oldItems, Is.EqualTo(slotItems));
@@ -167,9 +167,9 @@ namespace TheChest.Inventories.Tests.Slots.InventoryStackSlot
             var stackSize = this.random.Next(MIN_STACK_SIZE_TEST, MAX_STACK_SIZE_TEST);
             var slotItems = this.itemFactory.CreateMany(stackSize);
             var slot = this.slotFactory.WithItems(slotItems, stackSize);
+
             var replacingItems = this.itemFactory.CreateManyRandom(this.random.Next(1, stackSize));
             var expectedItems = (T[])replacingItems.Clone();
-
             slot.TryReplace(replacingItems, out _);
 
             Assert.That(slot.GetContents(), Is.EqualTo(expectedItems));