+++ CHANGELOG.md
@@ -14,6 +14,7 @@
   * `LazyStackInventory`
     * `TryAdd(T item, int amount)` - Tries to add the requested amount to the inventory and returns `true` only when all requested items are added
     * `TryAddAt(T item, int index, int amount)` - Tries to add the requested amount to a specific slot and returns `true` only when all requested items are added
+    * `TryReplace(T item, int index, int amount, out T[] oldItems)` - Tries to replace the requested amount to a specific slot and returns `true` only when the item is successfully replaced
 * New add methods with attempting to do an action without throwing exceptions when the action is not possible to be done because of the slot state. 
   * `InventorySlot`
     * `TryAdd(T item)` - Tries to add an item to the slot and returns `true` on success; otherwise `false`
@@ -23,6 +24,7 @@
     * `TryReplace(T[] items, out T[] oldItems)` - Tries to replace the items in the slot and returns `true` on success and outputs the replaced items; otherwise `false` and default value of `T[]` as the output
   * `InventoryLazyStackSlot`
     * `TryAdd(T item, int amount)` - Tries to add the requested amount to the slot and returns `true` only when all requested items are added
+    * `TryReplace(T item, int amount, out T[] oldItems)` - Tries to replace the requested amount to a specific slot and returns `true` only when the item is successfully replaced 
 
 ## What's Changed
 * `InventorySlot<T>.Add` method now throws `ArgumentNullException` when the param `item` is null instead of just returning false`
@@ -36,6 +38,7 @@
 
 ## What's Fixed
 * `CanAddItems` and `AddItems` from `LazyStackInventory<T>` now checks if the AvailableAmount in the slot is bigger than zero before trying to add items to it.
+* `StackInventory` methods that have `index` params doesn't throw `ArgumentOutOfException` anymore when searched index was equal to `Size`
 
 ## Known Issues
 * **No support for structs or value types**